@@ -1,3 +1,12 @@
 export const environment = {
-  production: true
+  production: true,
+  firebaseConfig : {
+    apiKey: "AIzaSyDA8FxeJRh6TwG1Qi8aAhld4e18xImkTKM",
+    authDomain: "dileepgaganr.firebaseapp.com",
+    databaseURL: "https://dileepgaganr.firebaseio.com",
+    projectId: "dileepgaganr",
+    storageBucket: "dileepgaganr.appspot.com",
+    messagingSenderId: "546460031975",
+    appId: "1:546460031975:web:9963ce5d19f6c4b31ebb42"
+  }
 };
