@@ -3,7 +3,16 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebaseConfig : {
+    apiKey: "AIzaSyDA8FxeJRh6TwG1Qi8aAhld4e18xImkTKM",
+    authDomain: "dileepgaganr.firebaseapp.com",
+    databaseURL: "https://dileepgaganr.firebaseio.com",
+    projectId: "dileepgaganr",
+    storageBucket: "dileepgaganr.appspot.com",
+    messagingSenderId: "546460031975",
+    appId: "1:546460031975:web:9963ce5d19f6c4b31ebb42"
+  }
 };
 
 /*
